+++ components/EventSection.tsx
@@ -5,7 +5,7 @@ import CalendarWeekly from "@/components/calendar/CalendarWeekly"
 import CalendarMonth from "@/components/calendar/CalendarMonth"
 import UpcomingEvents from "@/components/calendar/UpcomingEvents"
 import { SectionHeader } from "@/components/SectionHeader"
-import ButtonLink from "@/components/ui/ButtonLink"
+import ButtonLink from "@/components/button/ButtonLink"
 import { FaCalendarAlt } from "react-icons/fa"
 
 export interface DataProps {
