+++ mdx-components.tsx
@@ -1,7 +1,7 @@
 import type { MDXComponents } from "mdx/types"
 import Image from "next/image"
 import { ContentSection } from "@/components/ContentSection"
-import { ButtonLink } from "@/components/ui/ButtonLink"
+import { ButtonLink } from "@/components/button/ButtonLink"
 import {
   FeaturedCarousel,
   FeaturedCarouselItem,
